pyproject.toml
[tool.poetry]
name = "fastapi-with-strawberry"
version = "0.1.0"
description = ""
authors = ["abhi904485 <Abhishek9044855265@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql = {extras = ["fastapi"], version = "^0.256.1"}
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
